:sparkles: assembly-to-elf generation now supports labels/sections
The parser can now work out labels and symbols to their respective
addresses (as generated by the parser as well)

There's now support for Text, Data and BSS sections.

While this is a major step, there are still some problems to solve:
* Referecing variables (labels) declared in the data section from within
  the Text section can be somewhat problematic. It should work when
  running code from within this program, but when writing an ELF file,
  those addresses will become invalid, because of how relocation/linking
  works.

This problem could be surpassed by not resolving addresses/symbols when
writing ELF files, but this still has to be explored

The program can't still read the Data section from an ELF file and load
that into memory. There's probably going to be a problem related to
addressing of symbols as well for the reading of ELF files, but that
still hasn't been tackled yet.