:sparkles: gdb stub fully functional, which means the debugger is working \o/
instructions on how to use the debugger are going to be added soon!

the external compilation of assembly code (through make) now uses the
canonical representation of bytes, which is more useful to understand
how the endianness plays its role

elf.rs was deprecated and removed in favor of elreader.rs and
elfwriter.rs

memory.rs now requires an endianness to be set during its instantiation.
This endianness is used to store data for write operations. When
retrieving data from the memory, the caller can choose the endianness of
the resulting data. This allows other emulated components to handle data
using a different endianness than that of the memory itself. In turn,
this enables components working with different endianness to use the
same memory interface!

elfwriter.rs and elfreader.rs now use custom error types. In the future,
more custom error types are going to be used throughout this project,
and this is a nice start

many tests were rewritten to work with the new way memory works (with
predefined endianness)