:sparkles: support for 'hello world' programs on the way
The problem of using variables declared in the data section (from within
the text section) has been locally solved, and a general solution is
under way. Right now the solution is hard coded for the example found at
main.rs, but it should work for other programs as well.

The idea is to replicate what the 'magic' method does (a temporary method
created while testing a simple hello world program) whenever a symbol
out of the text section is used in the text section.

The solution includes adding a 'relocation' whenever a symbol declared
in the data section is referred to in the text section. This allows the
linker to regenerate the address of those symbols during the linking
stage, which in turn produces valid ELF files that can be run in a
riscv32 VM or in a GDB debugging session.